fixup! build: Generate `toolchain.cmake` in depends
To detect cross-compiling, the host and build platforms are compared.
The `build` variable is always an output of `config.sub`, but the `host`
is not. This can lead to false results. For example, on OpenBSD:
 - host=amd64-unknown-openbsd7.5
 - build=x86_64-unknown-openbsd7.5

This change replaces the `host` variable with `canonical_host`, which is
also always an output of `config.sub`. Thus, `canonical_host` and
`build` are always properly comparable.