guix: Adjust for Qt 6
1. Do not set `C{PLUS}_INCLUDE_PATH` variables for Darwin builds

When cross-building for macOS, the `C{PLUS}_INCLUDE_PATH` environment
variables modify the default include directories for both native GCC and
cross Clang compilers, which is overkill and undesirable for the latter.

This change avoids setting the `C{PLUS}_INCLUDE_PATH` environment
variables and instead sets the required `-isystem` flags for the native
compiler explicitly.

2. Add `ninja` package.

3. Adjust allowed symbol lists.