depends: Use base system's `sha256sum` utility
On FreeBSD, the `shasum` utility is provided by the `perl5` port, which
is not part of the base system and must be installed separately.
Note that this requirement is currently not documented in
`depends/README.md`.

This change switches to using the `sha256sum` utility, which is included
in the base system.

The same reasoning applies to Darwin, which is based on FreeBSD in this
context.