a fix for message:
```
The current fuzz target used the global random state.

This is acceptable, but requires the fuzz target to call
SeedRandomStateForTest(SeedRand::ZEROS) in the first line
of the FUZZ_TARGET function.

An alternative solution would be to avoid any use of globals.

Without a solution, fuzz instability and non-determinism can lead
to non-reproducible bugs or inefficient fuzzing.
```